Added reactive forms to customer detail

+++ src/app/dyn-customers/dyn-customer-detail/dyn-customer-detail.component.ts
@@ -1,6 +1,8 @@
 import { Component, Input } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
+import { FormBuilder, FormGroup } from '@angular/forms';
+
 import 'rxjs/add/operator/switchMap';
 
 import { Customer } from '../shared/dyn-customer.model';
@@ -17,35 +19,51 @@ export class CustomerDetailComponent {
   customer: Customer;
   customerId: String;
   result: any;
+  customerForm: FormGroup;
 
   constructor(
     private route: ActivatedRoute,
     private router: Router,
     private service: CustomerService,
+    private fb: FormBuilder,
     public dialogService: DynDialogService
-  ) {}
-  
+  ) { }
+
   ngOnInit() {
-    
+    this.createForm();
+
     this.route.params
       .switchMap((params: Params) => this.service.getCustomer(params['id']))
-      .subscribe((customer: Customer) => { 
+      .subscribe((customer: Customer) => {
         this.customer = customer[0]
+
+        this.customerForm.setValue({
+          name: this.customer.name
+        });
       });
   }
 
-  saveCustomer(){
+  createForm() {
+    this.customerForm = this.fb.group({
+      name: ''
+    });
+  }
+
+  saveCustomer() {
+    const formModel = this.customerForm.value;
+    this.customer.name = formModel.name as string;
     this.service.saveCustomer(this.customer);
     this.router.navigate(['/customers']);
   }
 
-  deleteCustomer(){
-    this.dialogService.confirm('Confirm Delete', 'Delete this customer, are you sure?') 
-    .subscribe(result => {
-      if (result) {
-        this.service.deleteCustomer(this.customer);
-        this.router.navigate(['/customers']);
-      }
-    });
+
+  deleteCustomer() {
+    this.dialogService.confirm('Confirm Delete', 'Delete this customer, are you sure?')
+      .subscribe(result => {
+        if (result) {
+          this.service.deleteCustomer(this.customer);
+          this.router.navigate(['/customers']);
+        }
+      });
   }
 }
